Fix docId/entryId type: Postgres SERIAL is INTEGER, not BIGINT
Task 2's reviewer built a standalone Hibernate/Testcontainers harness
against the real infra/seed/schema.sql and reproduced a schema
validation failure: EvidenceDocumentEntity.docId and
AuditEntryEntity.entryId were mapped as Long/BIGINT, but SERIAL PRIMARY
KEY in Postgres is a 4-byte INTEGER (BIGSERIAL would be BIGINT). This
would have broken Task 4's Testcontainers test with
"wrong column type ... found [serial], but expecting [bigint]" at
Spring context startup.

Changed both entity fields (and getters/setters) to Integer, and their
repositories' JpaRepository<..., Integer> generic to match. Also
corrected the same Long references in the plan document's Task 3/4
code (not yet executed) so the bug doesn't recur when those tasks run.

+++ case-system-mcp/src/main/java/com/workbench/mcp/entity/AuditEntryEntity.java
@@ -16,7 +16,7 @@ public class AuditEntryEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "entry_id")
-    private Long entryId;
+    private Integer entryId;
 
     @Column(name = "case_id")
     private String caseId;
@@ -30,11 +30,11 @@ public class AuditEntryEntity {
     @Column(name = "performed_at")
     private OffsetDateTime performedAt;
 
-    public Long getEntryId() {
+    public Integer getEntryId() {
         return entryId;
     }
 
-    public void setEntryId(Long entryId) {
+    public void setEntryId(Integer entryId) {
         this.entryId = entryId;
     }
