+++ case-system-mcp/src/main/java/com/workbench/mcp/repository/AuditEntryRepository.java
@@ -3,5 +3,5 @@ package com.workbench.mcp.repository;
 import com.workbench.mcp.entity.AuditEntryEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuditEntryRepository extends JpaRepository<AuditEntryEntity, Long> {
+public interface AuditEntryRepository extends JpaRepository<AuditEntryEntity, Integer> {
 }
