+++ case-system-mcp/src/main/java/com/workbench/mcp/repository/EvidenceDocumentRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface EvidenceDocumentRepository extends JpaRepository<EvidenceDocumentEntity, Long> {
+public interface EvidenceDocumentRepository extends JpaRepository<EvidenceDocumentEntity, Integer> {
 
     List<EvidenceDocumentEntity> findByCaseId(String caseId);
 }
